Calculates properly one-digit recurring patterns.

+++ Calculator/Pattern.cs
@@ -8,10 +8,10 @@ namespace Calculator
 {
     public class Pattern
     {
-        public byte IndexOf { get; set; }
+        public int IndexOf { get; set; }
         public string Value { get; set; }
 
-        public Pattern(string value, byte position=0)
+        public Pattern(string value, int position=0)
         {
             this.IndexOf = position;
             this.Value = value;
